ultimas_pruebas/obtener_propiedades_tablas_g/requirements.txt
functions-framework
google.cloud